+++ packages/data-context/src/sources/migration/codegen.ts
@@ -535,6 +535,7 @@ export function getSpecPattern (cfg: LegacyCypressConfigJson, testingType: Testi
 function formatWithBundledBabel (config: string) {
   const ast = parse(config)
 
+  // @ts-ignore - transitive babel types have a minor conflict - readonly vs non readonly.
   let { code } = generate(ast, {}, config)
   // By default babel generates imports like this:
   // const {