+++ scripts/github-actions/update-browser-versions.js
@@ -110,22 +110,9 @@ const updatePRTitle = async ({ context, github, baseBranch, branchName, descript
   })
 }
 
-const createPullRequest = async ({ context, github, baseBranch, branchName, description }) => {
-  await github.pulls.create({
-    owner: context.repo.owner,
-    repo: context.repo.repo,
-    base: baseBranch,
-    head: branchName,
-    title: `chore: ${description}`,
-    body: 'This PR was auto-generated to update the version(s) of Chrome for driver tests',
-    maintainer_can_modify: true,
-  })
-}
-
 module.exports = {
   getVersions,
   checkNeedForBranchUpdate,
   updateBrowserVersionsFile,
   updatePRTitle,
-  createPullRequest,
 }